requirements.txt
aiogram
# aioredis
environs
# SQLAlchemy
# asyncpg
# alembic
loguru

gspread==4.0.1
oauth2client~=4.1.3
google-api-python-client==2.29.0
google-auth-httplib2
google-auth-oauthlib
gspread-asyncio==1.5.2

